+++ src/main/java/insane96mcp/enhancedai/data/mpr/condition/EAIConditionsRegistry.java
@@ -1,0 +1,17 @@
+package insane96mcp.enhancedai.data.mpr.condition;
+
+import insane96mcp.enhancedai.EnhancedAI;
+import insane96mcp.mobspropertiesrandomness.data.json.condition.MPRCondition;
+
+import static insane96mcp.mobspropertiesrandomness.data.json.condition.ConditionsRegistry.CONDITIONS;
+
+public class EAIConditionsRegistry {
+	/// Use you own namespace
+	private static void register(String id, Class<? extends MPRCondition> clazz) {
+		CONDITIONS.put(EnhancedAI.location(id), clazz);
+	}
+
+	public static void init() {
+		register("is_leader", EAIIsLeaderCondition.class);
+	}
+}