+++ src/main/java/insane96mcp/enhancedai/data/mpr/condition/EAIIsLeaderCondition.java
@@ -1,0 +1,31 @@
+package insane96mcp.enhancedai.data.mpr.condition;
+
+import com.google.gson.*;
+import com.google.gson.annotations.JsonAdapter;
+import insane96mcp.enhancedai.modules.mobs.Leaders;
+import insane96mcp.mobspropertiesrandomness.data.json.condition.MPRCondition;
+import net.minecraft.world.entity.LivingEntity;
+
+import java.lang.reflect.Type;
+
+@JsonAdapter(EAIIsLeaderCondition.Serializer.class)
+public class EAIIsLeaderCondition extends MPRCondition {
+    public EAIIsLeaderCondition(boolean inverted) {
+        super(inverted);
+    }
+
+    protected boolean conditionCheck(LivingEntity living) {
+        return Leaders.isLeader(living);
+    }
+
+    public static class Serializer implements JsonDeserializer<EAIIsLeaderCondition>, JsonSerializer<EAIIsLeaderCondition> {
+        public EAIIsLeaderCondition deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
+            JsonObject jObject = json.getAsJsonObject();
+            return new EAIIsLeaderCondition(MPRCondition.deserializeInverted(jObject));
+        }
+
+        public JsonElement serialize(EAIIsLeaderCondition src, Type typeOfSrc, JsonSerializationContext context) {
+            return src.endSerialization(new JsonObject());
+        }
+    }
+}