+++ changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## Upcoming
+* Added Leader mobs
+  * Zombies and skeletons now have a 5% chance to spawn as a leader
+  * Leader mobs have heavily increased health and high armor and have 100% chance to spawn reinforcements when hit
+  * Each time they're hit, the chance decreases
+  * This feature has also an MPR condition `enhancedai:is_leader` to check if the mob is a leader.
 * Fisher mobs
   * Pulling items from inventory no longer launches them into the farlands